add grace period and jit

+++ metrics/gini.py
@@ -3,7 +3,7 @@ import numpy as np
 
 @jit(nopython=True)
 def gini(class_freq, total=None):
-    g = np.float(1.)
+    g = 1.
     if total is None:
         total = np.sum(class_freq)
     for k in class_freq:
